2.3.4

+++ src/org/bonitasoft/truckmilk/plugin/MilkPurgeArchive.java
@@ -7,11 +7,8 @@ import java.util.logging.Logger;
 
 import org.bonitasoft.engine.api.APIAccessor;
 import org.bonitasoft.engine.api.ProcessAPI;
-import org.bonitasoft.engine.bpm.process.ActivationState;
 import org.bonitasoft.engine.bpm.process.ArchivedProcessInstance;
 import org.bonitasoft.engine.bpm.process.ArchivedProcessInstancesSearchDescriptor;
-import org.bonitasoft.engine.bpm.process.ProcessDeploymentInfo;
-import org.bonitasoft.engine.bpm.process.ProcessDeploymentInfoSearchDescriptor;
 import org.bonitasoft.engine.exception.DeletionException;
 import org.bonitasoft.engine.exception.SearchException;
 import org.bonitasoft.engine.search.Order;
@@ -19,37 +16,32 @@ import org.bonitasoft.engine.search.SearchOptionsBuilder;
 import org.bonitasoft.engine.search.SearchResult;
 import org.bonitasoft.log.event.BEvent;
 import org.bonitasoft.log.event.BEvent.Level;
+import org.bonitasoft.log.event.BEventFactory;
 import org.bonitasoft.truckmilk.engine.MilkPlugIn;
+import org.bonitasoft.truckmilk.engine.MilkPlugInToolbox;
+import org.bonitasoft.truckmilk.engine.MilkPlugInToolbox.DelayResult;
+import org.bonitasoft.truckmilk.engine.MilkPlugInToolbox.ListProcessesResult;
 import org.bonitasoft.truckmilk.job.MilkJobExecution;
+import org.bonitasoft.truckmilk.toolbox.TypesCast;
 
 public class MilkPurgeArchive extends MilkPlugIn {
 
     static Logger logger = Logger.getLogger(MilkPurgeArchive.class.getName());
 
-    private static BEvent eventNoProcessMatchFilter = new BEvent(MilkPurgeArchive.class.getName(), 1,
-            Level.APPLICATIONERROR,
-            "No process match filter", "No process is found with the given filter", "This filter does not apply.",
-            "Check the process name");
 
-    private static BEvent eventNoProcessForFilter = new BEvent(MilkPurgeArchive.class.getName(), 2,
-            Level.APPLICATIONERROR,
-            "Filter is not active", "No processes was found for all the filter, search will not run",
-            "No filter at all apply, assuming configuration want to apply only on some process",
-            "Check the process name");
-
-    private static BEvent eventDeletionSuccess = new BEvent(MilkPurgeArchive.class.getName(), 3, Level.SUCCESS,
+    private static BEvent eventDeletionSuccess = new BEvent(MilkPurgeArchive.class.getName(), 1, Level.SUCCESS,
             "Deletion done with success", "Archived Cases are deleted with success");
 
-    private static BEvent eventDeletionFailed = new BEvent(MilkPurgeArchive.class.getName(), 4, Level.ERROR,
+    private static BEvent eventDeletionFailed = new BEvent(MilkPurgeArchive.class.getName(), 2, Level.ERROR,
             "Error during deletion", "An error arrived during the deletion of archived cases", "Cases are not deleted", "Check the exception");
 
-    private static BEvent eventSearchFailed = new BEvent(MilkPurgeArchive.class.getName(), 5, Level.ERROR,
+    private static BEvent eventSearchFailed = new BEvent(MilkPurgeArchive.class.getName(), 3, Level.ERROR,
             "Search failed", "Search failed task return an error", "No retry can be performed", "Check the error");
 
-    private static PlugInParameter cstParamDelayInDay = PlugInParameter.createInstance("delayinday", "Delay in day", TypeParameter.LONG, 10L, "The case must be older than this number, in days. 0 means all archived case is immediately in the perimeter");
+    private static PlugInParameter cstParamDelayInDay = PlugInParameter.createInstance("delayinday", "Delay in day", TypeParameter.DELAY, MilkPlugInToolbox.DELAYSCOPE.MONTH + ":3", "The case must be older than this number, in days. 0 means all archived case is immediately in the perimeter");
     private static PlugInParameter cstParamMaximumDeletionInCases = PlugInParameter.createInstance("maximumdeletionincase", "Maximum deletion in cases", TypeParameter.LONG, 10L, "Maximum cases deleted in one execution, to not overload the engine. Maximum hard coded at 5000.");
     private static PlugInParameter cstParamMaximumDeletionInMinutes = PlugInParameter.createInstance("maximumdeletioninminutes", "Maximum deletion in minutes", TypeParameter.LONG, 3L, "Maximum time in minutes for the job. After this time, it will stop.");
-    private static PlugInParameter cstParamProcessfilter = PlugInParameter.createInstance("processfilter", "Process Filter", TypeParameter.ARRAY, null, "Give a list of process name. Name must be exact, no version is given (all versions will be purged)");
+    private static PlugInParameter cstParamProcessFilter = PlugInParameter.createInstance("processfilter", "Process Filter", TypeParameter.ARRAYPROCESSNAME, null, "Give a list of process name. Name must be exact, no version is given (all versions will be purged)");
 
     public MilkPurgeArchive() {
         super(TYPE_PLUGIN.EMBEDED);
@@ -69,74 +61,47 @@ public class MilkPurgeArchive extends MilkPlugIn {
         return new ArrayList<>();
     }
 
-    @SuppressWarnings("unchecked")
-    @Override
-    public PlugTourOutput execute(MilkJobExecution milkPlugInExecution, APIAccessor apiAccessor) {
-        PlugTourOutput plugTourOutput = milkPlugInExecution.getPlugTourOutput();
 
+    @Override
+    public PlugTourOutput execute(MilkJobExecution jobExecution, APIAccessor apiAccessor) {
+        PlugTourOutput plugTourOutput = jobExecution.getPlugTourOutput();
+
+        long maximumDeletionInCases = jobExecution.getInputLongParameter(cstParamMaximumDeletionInCases);
+        if (maximumDeletionInCases > 10000)
+            maximumDeletionInCases = 10000;
         ProcessAPI processAPI = apiAccessor.getProcessAPI();
         // get Input 
-        List<String> listProcessName = (List<String>) milkPlugInExecution.getInputListParameter(cstParamProcessfilter);
-        Long delayInDay = milkPlugInExecution.getInputLongParameter(cstParamDelayInDay);
 
-        milkPlugInExecution.setPleaseStopAfterTime(milkPlugInExecution.getInputLongParameter(cstParamMaximumDeletionInMinutes), 24 * 60L);
-        milkPlugInExecution.setPleaseStopAfterManagedItems(milkPlugInExecution.getInputLongParameter(cstParamMaximumDeletionInCases), 5000L);
+        SearchOptionsBuilder searchActBuilder = new SearchOptionsBuilder(0, (int) maximumDeletionInCases + 1);
+        SearchResult<ArchivedProcessInstance> searchArchivedProcessInstance;
 
-        List<Long> listProcessDefinitionId = new ArrayList<>();
+        try {
+            ListProcessesResult listProcessResult = MilkPlugInToolbox.completeListProcess(jobExecution, cstParamProcessFilter, true, searchActBuilder, ArchivedProcessInstancesSearchDescriptor.PROCESS_DEFINITION_ID, apiAccessor.getProcessAPI());
 
-        // Filter on process?
-        SearchOptionsBuilder searchActBuilder = new SearchOptionsBuilder(0, milkPlugInExecution.getPleaseStopAfterManagerItems().intValue());
-
-        if (listProcessName != null && !listProcessName.isEmpty()) {
-
-            for (String processName : listProcessName) {
-                SearchOptionsBuilder searchOptionBuilder = new SearchOptionsBuilder(0, 1000);
-                searchOptionBuilder.filter(ProcessDeploymentInfoSearchDescriptor.NAME, processName);
-                searchOptionBuilder.filter(ProcessDeploymentInfoSearchDescriptor.ACTIVATION_STATE,
-                        ActivationState.ENABLED.name());
-                try {
-                    SearchResult<ProcessDeploymentInfo> searchProcessDeployment = processAPI
-                            .searchProcessDeploymentInfos(searchOptionBuilder.done());
-                    if (searchProcessDeployment.getCount() == 0) {
-                        plugTourOutput.addEvent(new BEvent(eventNoProcessMatchFilter, "Filter[" + processName + "]"));
-
-                    }
-                    for (ProcessDeploymentInfo processInfo : searchProcessDeployment.getResult()) {
-                        listProcessDefinitionId.add(processInfo.getProcessId());
-                    }
-                } catch (SearchException e1) {
-                    plugTourOutput.addEvent(new BEvent(eventSearchFailed, e1, ""));
-                    plugTourOutput.executionStatus = ExecutionStatus.ERROR;
-                    return plugTourOutput;
-                }
-            }
-            if (!listProcessDefinitionId.isEmpty()) {
+            if (listProcessResult.listProcessDeploymentInfo.isEmpty()) {
                 // filter given, no process found : stop now
-                plugTourOutput.addEvent(eventNoProcessForFilter);
+                plugTourOutput.addEvents(listProcessResult.listEvents);
                 plugTourOutput.executionStatus = ExecutionStatus.BADCONFIGURATION;
                 return plugTourOutput;
             }
-            searchActBuilder.leftParenthesis();
-            for (int i = 0; i < listProcessDefinitionId.size(); i++) {
-                if (i > 0)
-                    searchActBuilder.or();
 
-                searchActBuilder.filter(ArchivedProcessInstancesSearchDescriptor.PROCESS_DEFINITION_ID,
-                        listProcessDefinitionId.get(i));
+            DelayResult delayResult = MilkPlugInToolbox.getTimeFromDelay(jobExecution, cstParamDelayInDay, new Date(), false);
+            if (BEventFactory.isError(delayResult.listEvents)) {
+                plugTourOutput.addEvents(delayResult.listEvents);
+                plugTourOutput.executionStatus = ExecutionStatus.ERROR;
+                return plugTourOutput;
             }
-            searchActBuilder.rightParenthesis();
-            searchActBuilder.and();
+            long timeSearch = delayResult.delayDate.getTime();
 
-        } // end list process name
+            jobExecution.setPleaseStopAfterTime(jobExecution.getInputLongParameter(cstParamMaximumDeletionInMinutes), 24 * 60L);
+            jobExecution.setPleaseStopAfterManagedItems(maximumDeletionInCases, 10000L);
 
-        // Now, search items to delete
-        Date currentDate = new Date();
-        long timeSearch = currentDate.getTime() - delayInDay * 1000 * 60 * 60 * 24;
+            // Now, search items to delete
 
-        searchActBuilder.lessOrEquals(ArchivedProcessInstancesSearchDescriptor.ARCHIVE_DATE, timeSearch);
-        searchActBuilder.sort(ArchivedProcessInstancesSearchDescriptor.ARCHIVE_DATE, Order.ASC);
-        SearchResult<ArchivedProcessInstance> searchArchivedProcessInstance;
-        try {
+         
+            searchActBuilder.lessOrEquals(ArchivedProcessInstancesSearchDescriptor.ARCHIVE_DATE, timeSearch);
+            searchActBuilder.sort(ArchivedProcessInstancesSearchDescriptor.ARCHIVE_DATE, Order.ASC);
+
             searchArchivedProcessInstance = processAPI.searchArchivedProcessInstances(searchActBuilder.done());
             if (searchArchivedProcessInstance.getCount() == 0) {
                 plugTourOutput.executionStatus = ExecutionStatus.SUCCESSNOTHING;
@@ -148,16 +113,31 @@ public class MilkPurgeArchive extends MilkPlugIn {
             plugTourOutput.executionStatus = ExecutionStatus.ERROR;
             return plugTourOutput;
         }
+        
+        
+        // do the purge now
+        Long beginTime = System.currentTimeMillis();
         List<Long> sourceProcessInstanceIds = new ArrayList<>();
-
-        for (ArchivedProcessInstance archivedProcessInstance : searchArchivedProcessInstance.getResult()) {
-
-            sourceProcessInstanceIds.add(archivedProcessInstance.getSourceObjectId());
-        }
         try {
+        for (ArchivedProcessInstance archivedProcessInstance : searchArchivedProcessInstance.getResult()) {
+            // proceed page per page
+            sourceProcessInstanceIds.add(archivedProcessInstance.getSourceObjectId());
+            if (sourceProcessInstanceIds.size() == 50)
+            {
+                processAPI.deleteArchivedProcessInstancesInAllStates(sourceProcessInstanceIds);
+                plugTourOutput.nbItemsProcessed += sourceProcessInstanceIds.size();
+                sourceProcessInstanceIds.clear();
+            }
+        }
+        // reliquat
+        if (! sourceProcessInstanceIds.isEmpty())
+        {
             processAPI.deleteArchivedProcessInstancesInAllStates(sourceProcessInstanceIds);
             plugTourOutput.nbItemsProcessed += sourceProcessInstanceIds.size();
-            plugTourOutput.addEvent(new BEvent(eventDeletionSuccess, "Purge:" + sourceProcessInstanceIds.size()));
+        }        
+        Long endTime = System.currentTimeMillis();
+        
+        plugTourOutput.addEvent(new BEvent(eventDeletionSuccess, "Purge:" + plugTourOutput.nbItemsProcessed+" in "+TypesCast.getHumanDuration(endTime-beginTime, true)));
 
         } catch (DeletionException e) {
             logger.severe("Error Delete Archived ProcessInstance=[" + sourceProcessInstanceIds + "] Error[" + e.getMessage() + "]");
@@ -178,7 +158,7 @@ public class MilkPurgeArchive extends MilkPlugIn {
         plugInDescription.addParameter(cstParamDelayInDay);
         plugInDescription.addParameter(cstParamMaximumDeletionInCases);
         plugInDescription.addParameter(cstParamMaximumDeletionInMinutes);
-        plugInDescription.addParameter(cstParamProcessfilter);
+        plugInDescription.addParameter(cstParamProcessFilter);
         return plugInDescription;
     }
 }
